# routes/api 2/openapi.yaml
openapi: 3.0.3
info:
  title: Coinet API
  version: 1.0.0
  description: >
    Modular, extensible, and production-grade API for Coinet, inspired by Apple, Canva, TradingView, and Solana.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Portfolio:
      type: object
      properties:
        id: { type: string }
        user_id: { type: string }
        name: { type: string }
        holdings: { type: array, items: { $ref: '#/components/schemas/Holding' } }
    Holding:
      type: object
      properties:
        id: { type: string }
        portfolio_id: { type: string }
        symbol: { type: string }
        quantity: { type: number }
        avg_cost: { type: number }
...
security:
  - bearerAuth: []
paths:
...
  /portfolios:
    post:
      tags: [Portfolios]
      summary: Create portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Portfolio' } } } }
    get:
      tags: [Portfolios]
      summary: List user portfolios
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Portfolio' } } } } }
  /portfolios/{id}:
    get:
      tags: [Portfolios]
      summary: Get portfolio details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Portfolio' } } } }
    put:
      tags: [Portfolios]
      summary: Update (rename) portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Portfolio' } } } }
    delete:
      tags: [Portfolios]
      summary: Delete portfolio
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /portfolios/{id}/holdings:
    post:
      tags: [Portfolios]
      summary: Add/update holding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Holding'
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Holding' } } } }
  /portfolios/{id}/holdings/{holdingId}:
    delete:
      tags: [Portfolios]
      summary: Remove a holding
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: holdingId
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...